Move all cdp tests into src
