+++ plugin-server/src/cdp/services/hog-executor.service.test.ts
@@ -1,14 +1,15 @@
 import { DateTime } from 'luxon'
 
+import { truth } from '~/tests/helpers/truth'
+
 import { HogExecutorService } from '../../../src/cdp/services/hog-executor.service'
 import { HogFunctionManagerService } from '../../../src/cdp/services/hog-function-manager.service'
 import { HogFunctionInvocation, HogFunctionType } from '../../../src/cdp/types'
 import { Hub } from '../../../src/types'
 import { createHub } from '../../../src/utils/db/hub'
 import { status } from '../../../src/utils/status'
-import { truth } from '../../helpers/truth'
-import { HOG_EXAMPLES, HOG_FILTERS_EXAMPLES, HOG_INPUTS_EXAMPLES } from '../examples'
-import { createHogExecutionGlobals, createHogFunction, createInvocation } from '../fixtures'
+import { HOG_EXAMPLES, HOG_FILTERS_EXAMPLES, HOG_INPUTS_EXAMPLES } from '../_tests/examples'
+import { createHogExecutionGlobals, createHogFunction, createInvocation } from '../_tests/fixtures'
 
 jest.mock('../../../src/utils/status', () => ({
     status: {
