+++ plugin-server/src/cdp/services/legacy-plugin-executor.service.test.ts
@@ -1,17 +1,17 @@
 import { DateTime } from 'luxon'
 
-import {
-    createHogExecutionGlobals,
-    createHogFunction,
-    createInvocation,
-    insertHogFunction as _insertHogFunction,
-} from '~/tests/cdp/fixtures'
 import { forSnapshot } from '~/tests/helpers/snapshots'
 import { getFirstTeam, resetTestDatabase } from '~/tests/helpers/sql'
 
 import { createPlugin, createPluginConfig } from '../../../tests/helpers/sql'
 import { Hub, PluginConfig, Team } from '../../types'
 import { closeHub, createHub } from '../../utils/db/hub'
+import {
+    createHogExecutionGlobals,
+    createHogFunction,
+    createInvocation,
+    insertHogFunction as _insertHogFunction,
+} from '../_tests/fixtures'
 import { DESTINATION_PLUGINS_BY_ID, TRANSFORMATION_PLUGINS_BY_ID } from '../legacy-plugins'
 import { LegacyDestinationPlugin, LegacyTransformationPlugin } from '../legacy-plugins/types'
 import { HogFunctionInvocation, HogFunctionInvocationGlobalsWithInputs, HogFunctionType } from '../types'
